chore: rename utils to podspec
Problem: the "utils" package in fluence does not describe
what it does, and these functions are hard for the developer
to find.
Solution: rename to podspec to reflect this is what the
package is oriented to do.
