+++ sig-scheduler-plugins/pkg/fluence/podspec/podspec.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package utils
+package podspec
 
 import (
 	"strings"
